replacing WEAVE with NUMBA: calculation of csm included
'faverage' is replaced by 'calcCSM'. In order to reduce computational cost by approx 0.7, only the upper triangular matrix of the CSM is calculated during the ensemble average. The whole CSM is produced by complex transpose of the averaged upper CSM. 
Furthermore: In earlier versions the CSM was stored as its transpose, which is now fixed. This means that iwith this commit the beamformers should produce wrong results.